move simple_task_queue
